demos/02_lcel_chain/requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.3.0
python-dotenv==1.2.1
